mainapp/pyproject.toml
[tool.poetry]
name = "stories"
version = "0.3.0"
description = ""
authors = ["ourteam"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = {extras = ["all"], version = "^0.61.2"}
emails = "^0.6"
pyjwt = "^1.7.1"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
psycopg2-binary = "^2.8.5"
Authlib = "^0.14.3"
httpx = "^0.14.0"
typer = "^0.3.2"
tortoise-orm = "^0.16.18"
aerich = "^0.3.3"
asyncpg = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"